Add files via upload
This code defines a Searching class that allows a user to input student test marks and names until the user enter -1. It then displays the names and marks in a tabular format, computes the average mark and identifies the highest and lowest marks along with the corresponding students. 
The user can subsequently search for a student's name to retrieve their test mark. 
The program utilizes arrays to store the marks and names, employing loops for data collection and processing. 
Overall, it combines user input handling, basic statistics, and searching functionalities within an interactive console application.